chore: split ImageAppContext into focused modules
- Extract createDebouncedTask and scheduleIdleTask to src/lib/asyncUtils.ts
- Extract form control state to useImageFormState SolidJS hook
- Extract processing pipeline to src/services/imageProcessingPipeline.ts
- Move AppState/AppActions interfaces to src/components/app/state/appTypes.ts
- ImageAppContext reduced from ~580 to 188 lines
- useImageApp() public API unchanged

Closes #107
